Added authorization and rate limiting

+++ api/comments.js
@@ -1,6 +1,8 @@
 const router = require('express').Router();
 
 const { validateAgainstSchema } = require('../lib/validation');
+const { generateAuthToken, requireAuthentication} = require('../lib/auth');
+const mysqlPool = require('../lib/mysqlPool');
 const {
     CommentSchema,
     getCommentsPage,
@@ -97,12 +99,14 @@ router.get('/:id', async (req, res, next) => {
 /*
  * Route to replace data for a comment.
  */
-router.put('/:id', async (req, res, next) => {
-  const id = parseInt(req.params.id)
-  if (id === req.body.commentId){
-    if (validateAgainstSchema(req.body, CommentSchema)) {
+router.put('/:id', requireAuthentication, async (req, res, next) => {
+  if (req.user != req.body.userId) {
+    res.status(403).send({
+      error: "Unauthorized to access the specified resource"
+    });
+  } else {
       try {
-        
+        const id = parseInt(req.params.id)
         const updateSuccessful = await replaceCommentById(id, req.body);
         if (updateSuccessful) {
           res.status(200).send({
@@ -119,15 +123,6 @@ router.put('/:id', async (req, res, next) => {
           error: "Unable to update specified comment.  Please try again later."
         });
       }
-    } else {
-      res.status(400).send({
-        error: "Request body is not a valid comment object"
-      });
-    }
-  }else{
-    res.status(500).send({
-      error: "Unable to update specified comment. Can not change the cooment id."
-    });
   }
 });
 
@@ -135,19 +130,29 @@ router.put('/:id', async (req, res, next) => {
 /*
 * Route to delete a comment.
 */
-router.delete('/:id', async (req, res, next) => {
-  try {
-    const deleteSuccessful = await deleteCommentById(parseInt(req.params.id));
-    if (deleteSuccessful) {
-      res.status(204).end();
-    } else {
-      next();
-    }
-  } catch (err) {
-    console.error(err);
-    res.status(500).send({
-      error: "Unable to delete comment.  Please try again later."
+router.delete('/:id', requireAuthentication, async (req, res, next) => {
+  const [ results ] = await mysqlPool.query(
+    'SELECT * FROM comments WHERE id = ?',
+    [ req.params.id]
+  );
+  if (req.user != results[0].userId) {
+    res.status(403).send({
+      error: "Unauthorized to access the specified resource"
     });
+  } else {
+    try {
+      const deleteSuccessful = await deleteCommentById(parseInt(req.params.id));
+      if (deleteSuccessful) {
+        res.status(204).end();
+      } else {
+        next();
+      }
+    } catch (err) {
+      console.error(err);
+      res.status(500).send({
+        error: "Unable to delete comment.  Please try again later."
+      });
+    }
   }
 });
